Revert package name
