+++ Demo/src/main/java/com/paypal/android/ui/paypalbuttons/ButtonFundingType.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.paypal
+package com.paypal.android.ui.paypalbuttons
 
 enum class ButtonFundingType {
     PAYPAL,
