+++ Demo/src/main/java/com/paypal/android/ui/paypalbuttons/PayPalCreditButtonColorOptionList.kt
@@ -1,22 +1,22 @@
-package com.paypal.android.ui.paypal
+package com.paypal.android.ui.paypalbuttons
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.paypal.android.R
-import com.paypal.android.paymentbuttons.PayPalButtonColor
+import com.paypal.android.paymentbuttons.PayPalCreditButtonColor
 import com.paypal.android.uishared.components.OptionList
 
 @Composable
-fun PayPalButtonColorOptionList(
-    selectedOption: PayPalButtonColor,
-    onSelection: (PayPalButtonColor) -> Unit
+fun PayPalCreditButtonColorOptionList(
+    selectedOption: PayPalCreditButtonColor,
+    onSelection: (PayPalCreditButtonColor) -> Unit
 ) {
     OptionList(
         title = stringResource(id = R.string.pay_pal_button_color),
-        options = PayPalButtonColor.entries.map { it.name },
+        options = PayPalCreditButtonColor.entries.map { it.name },
         selectedOption = selectedOption.name,
         onSelectedOptionChange = { option ->
-            onSelection(PayPalButtonColor.valueOf(option))
+            onSelection(PayPalCreditButtonColor.valueOf(option))
         }
     )
 }
